requirements.txt
pdf2image>=1.16.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0
Pillow>=9.5.0
imgkit>=1.2.3
requests>=2.31.0
python-magic>=0.4.27
python-magic-bin>=0.4.14
cryptography>=41.0.0
numpy>=1.24.0
opencv-python>=4.8.0
chardet>=5.2.0
tqdm>=4.66.0
colorama>=0.4.6
PyQt6>=6.5.0
PyQt6-WebEngine>=6.5.0 